Start #3 and #6

+++ src/course3/lcm.java
@@ -6,15 +6,15 @@ public class lcm {
         Scanner sc = new Scanner(System.in);
         int n1 = sc.nextInt();
         int n2 = sc.nextInt();
-        int lcm = lcm(n1, n2);
+        long lcm = lcm(n1, n2);
         System.out.println(lcm);
     }
 
-    private static int lcm(int n1, int n2) {
+    private static long lcm(int n1, int n2) {
         return (n1 * n2)/ gcd(n1, n2);
     }
 
-    private static int gcd(int n1, int n2) {
+    private static long gcd(int n1, int n2) {
         if (n2 == 0) {
             return n1;
         }
